# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_update_team_logo_tournaments__tournament_id__teams__team_id__logo_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_update_team_logo_tournaments__tournament_id__teams__team_id__logo_post",
        "type": "object"
      },
...
      "FullTeamWithPlayers": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "draws": {
            "default": 0,
            "title": "Draws",
            "type": "integer"
          },
          "elo_score": {
            "default": "1200",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Elo Score",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "logo_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Path"
          },
          "losses": {
            "default": 0,
            "title": "Losses",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/Player"
            },
            "title": "Players",
            "type": "array"
          },
          "swiss_score": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Swiss Score",
            "type": "string"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          },
          "wins": {
            "default": 0,
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "created",
          "name",
          "tournament_id",
          "active",
          "elo_score",
          "swiss_score",
          "wins",
          "draws",
          "losses",
          "logo_path",
          "id",
          "players"
        ],
        "title": "FullTeamWithPlayers",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedTeams": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/FullTeamWithPlayers"
            },
            "title": "Teams",
            "type": "array"
          }
        },
        "required": [
          "count",
          "teams"
        ],
        "title": "PaginatedTeams",
        "type": "object"
      },
      "Player": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "draws": {
            "default": 0,
            "title": "Draws",
            "type": "integer"
          },
          "elo_score": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Elo Score",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "losses": {
            "default": 0,
            "title": "Losses",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "swiss_score": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Swiss Score",
            "type": "string"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          },
          "wins": {
            "default": 0,
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "active",
          "name",
          "created",
          "tournament_id",
          "elo_score",
          "swiss_score",
          "wins",
          "draws",
          "losses",
          "id"
        ],
        "title": "Player",
        "type": "object"
      },
...
      "SingleTeamResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Team"
          }
        },
        "required": [
          "data"
        ],
        "title": "SingleTeamResponse",
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "SuccessResponse",
        "type": "object"
      },
...
      "Team": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "draws": {
            "default": 0,
            "title": "Draws",
            "type": "integer"
          },
          "elo_score": {
            "default": "1200",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Elo Score",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "logo_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Path"
          },
          "losses": {
            "default": 0,
            "title": "Losses",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "swiss_score": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Swiss Score",
            "type": "string"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          },
          "wins": {
            "default": 0,
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "created",
          "name",
          "tournament_id",
          "active",
          "elo_score",
          "swiss_score",
          "wins",
          "draws",
          "losses",
          "logo_path",
          "id"
        ],
        "title": "Team",
        "type": "object"
      },
      "TeamBody": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "name": {
            "maxLength": 30,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "player_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Player Ids",
            "type": "array",
            "uniqueItems": true
          }
        },
        "required": [
          "name",
          "active",
          "player_ids"
        ],
        "title": "TeamBody",
        "type": "object"
      },
      "TeamMultiBody": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "names": {
            "minLength": 1,
            "title": "Names",
            "type": "string"
          }
        },
        "required": [
          "names",
          "active"
        ],
        "title": "TeamMultiBody",
        "type": "object"
      },
      "TeamsWithPlayersResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaginatedTeams"
          }
        },
        "required": [
          "data"
        ],
        "title": "TeamsWithPlayersResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n### Description\nThis API allows you to do everything the frontend of [Casa Familia Padel Club](https://github.com/evroon/bracket)\nallows you to do (the frontend uses this API as well).\n\nFore more information, see the [documentation](https://docs.bracketapp.nl).\n\n### Table of Contents\n*(links only work for [ReDoc](https://api.bracketapp.nl/redoc), not for Swagger UI)*\n\n- [Auth](#tag/Auth)\n\n- [Clubs](#tag/Clubs)\n\n- [Courts](#tag/Courts)\n\n- [Internals](#tag/Internals)\n\n- [Matches](#tag/Matches)\n\n- [Players](#tag/Players)\n\n- [Rankings](#tag/Rankings)\n\n- [Rounds](#tag/Rounds)\n\n- [Stage Items](#tag/Stage-Items)\n\n- [Stage Item Inputs](#tag/Stage-Item-Inputs)\n\n- [Stages](#tag/Stages)\n\n- [Teams](#tag/Teams)\n\n- [Tournaments](#tag/Tournaments)\n\n- [Users](#tag/Users)\n\n### Links\nGitHub: <https://github.com/evroon/bracket>\n\nDocs: <https://docs.bracketapp.nl>\n\nDemo: <https://www.bracketapp.nl/demo>\n\nAPI docs (Redoc): <https://api.bracketapp.nl/redoc>\n\nAPI docs (Swagger UI): <https://api.bracketapp.nl/docs>\n",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "summary": "API for Casa Familia Padel Club, an open source tournament system.",
    "title": "Casa Familia Padel Club API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tournaments/{tournament_id}/teams": {
      "get": {
        "operationId": "get_teams_tournaments__tournament_id__teams_get",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "description": "Max number of results in a single page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Max number of results in a single page.",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Filter results starting from this offset.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Filter results starting from this offset.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort_direction",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Sort Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "name",
              "enum": [
                "name",
                "elo_score",
                "swiss_score",
                "wins",
                "draws",
                "losses",
                "active",
                "created"
              ],
              "title": "Sort By",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamsWithPlayersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Teams",
        "tags": [
          "Teams"
        ]
      },
      "post": {
        "operationId": "create_team_tournaments__tournament_id__teams_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleTeamResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/tournaments/{tournament_id}/teams/{team_id}": {
      "delete": {
        "operationId": "delete_team_tournaments__tournament_id__teams__team_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Team",
        "tags": [
          "Teams"
        ]
      },
      "put": {
        "operationId": "update_team_by_id_tournaments__tournament_id__teams__team_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleTeamResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Team By Id",
        "tags": [
          "Teams"
        ]
      }
    },
    "/tournaments/{tournament_id}/teams/{team_id}/logo": {
      "post": {
        "operationId": "update_team_logo_tournaments__tournament_id__teams__team_id__logo_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_team_logo_tournaments__tournament_id__teams__team_id__logo_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleTeamResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Team Logo",
        "tags": [
          "Teams"
        ]
      }
    },
    "/tournaments/{tournament_id}/teams_multi": {
      "post": {
        "operationId": "create_multiple_teams_tournaments__tournament_id__teams_multi_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamMultiBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Multiple Teams",
        "tags": [
          "Teams"
        ]
      }
    },
...
  }
}